requirements.txt
anthropic>=0.92.0
python-dotenv>=1.0.0
openai>=1.0.0
